+++ src/pages/Main.jsx
@@ -1,35 +1,81 @@
-import React from "react";
-import { useNavigate } from "react-router-dom";
+import React, { useState, useEffect } from "react";
+import { Link, useNavigate } from "react-router-dom";
 import styled from "styled-components";
 import Header from "../common/Header";
 import Container from "../common/Container";
 import { useSelector, useDispatch } from "react-redux";
 import { deletePost } from "../redux/slice/postSlice"; // index.js에서 deletePost 가져오기
-
+import { auth } from "../firebase"; // Firebase 모듈에서 auth 객체를 가져온다
 export default function Main() {
   // 리덕스 스토어의 '게시글' 상태를 조회
   const todos = useSelector((state) => state.게시글);
   const dispatch = useDispatch();
   const navigate = useNavigate();
+  const [user, setUser] = useState(null); // 로그인한 사용자 정보를 상태로 관리
 
-  const handleEditClick = (postId) => {
-    navigate(`/edit/${postId}`); // 수정 버튼 클릭 시 해당 게시물의 수정 페이지로 이동
+  // 로그인 상태 감지
+  useEffect(() => {
+    const unsubscribe = auth.onAuthStateChanged((user) => {
+      if (user) {
+        setUser(user); // 로그인한 사용자 정보 설정
+      } else {
+        setUser(null); // 로그아웃 시 사용자 정보 초기화
+      }
+    });
+
+    // 컴포넌트 언마운트 시에 이벤트 리스너 구독 해제
+    return () => unsubscribe();
+  }, []);
+
+  const handleEditClick = (post) => {
+    // 로그인 여부 확인
+    if (user) {
+      // 게시물의 author와 로그인한 이메일이 일치하는지 확인
+      if (post.author === user.email) {
+        navigate(`/edit/${post.id}`); // 로그인한 사용자와 게시물의 작성자가 일치하면 수정 페이지로 이동
+      } else {
+        // 일치하지 않는 경우 경고창 띄우기
+        alert("어허~남의 것을 탐하지말지어다.");
+      }
+    } else {
+      // 로그인이 안된 경우 경고창 띄우기
+      alert("로그인 후에 게시물을 수정할 수 있습니다.");
+    }
   };
 
   const handleCreateClick = () => {
-    navigate(`/create`); // 추가 버튼 클릭 시 새로운 게시물 생성 페이지로 이동
+    // 로그인 여부 확인
+    if (user) {
+      navigate("/create"); // 로그인 상태인 경우 추가 페이지로 이동
+    } else {
+      // 로그인이 안된 경우 경고창 띄우기
+      alert("로그인 후에 게시물을 추가할 수 있습니다.");
+    }
   };
 
   const handleDeleteClick = (postId) => {
-    // 확인 알림 창 띄우기
-    const confirmDelete = window.confirm(
-      "기억은 머리 속에서 살지만 추억은 가슴 속에서 산다. 정말로 삭제하시겠습니까?"
-    );
+    // 로그인 여부 확인
+    if (user) {
+      // 게시물의 author와 로그인한 이메일이 일치하는지 확인
+      const post = todos.find((post) => post.id === postId);
+      if (post && post.author === user.email) {
+        // 확인 알림 창 띄우기
+        const confirmDelete = window.confirm(
+          "기억은 머리 속에서 살지만 추억은 가슴 속에서 산다. 정말로 삭제하시겠습니까?"
+        );
 
-    if (confirmDelete) {
-      // 삭제 진행
-      dispatch(deletePost(postId)); // deletePost 액션 생성자 함수를 호출하여 액션 디스패치
-      navigate("/");
+        if (confirmDelete) {
+          // 삭제 진행
+          dispatch(deletePost(postId)); // deletePost 액션 생성자 함수를 호출하여 액션 디스패치
+          navigate("/");
+        }
+      } else {
+        // 일치하지 않는 경우 경고창 띄우기
+        alert("어허~남의 것을 탐하지말지어다.");
+      }
+    } else {
+      // 로그인이 안된 경우 경고창 띄우기
+      alert("로그인 후에 게시물을 삭제할 수 있습니다.");
     }
   };
 
@@ -52,7 +98,7 @@ export default function Main() {
             <AuthorName>{post.author}</AuthorName>
             <ActionButtons>
               <div>
-                <EditButton onClick={() => handleEditClick(post.id)}>
+                <EditButton onClick={() => handleEditClick(post)}>
                   수정
                 </EditButton>{" "}
                 {/* '수정' 버튼 */}
